Exception handling.

+++ GymServices/src/main/java/com/juancrud/gym/services/BaseService.java
@@ -2,14 +2,18 @@ package com.juancrud.gym.services;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Map;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.repository.CrudRepository;
 
+import com.juancrud.gym.services.exceptions.GymServiceException;
 import com.juancrud.gym.services.interfaces.IBaseService;
 import com.juancrud.gym.services.mappers.IModelEntityMapper;
+import com.juancrud.gym.services.models.EntityModel;
 
-public abstract class BaseService<M, E, ID> implements IBaseService<M, ID> {
+public abstract class BaseService<M extends EntityModel<ID>, E, ID> implements IBaseService<M, ID> {
 	
 	@Autowired
 	private CrudRepository<E, ID> repository;
@@ -25,22 +29,41 @@ public abstract class BaseService<M, E, ID> implements IBaseService<M, ID> {
 		return result;
 	}
 	
+	public Collection<M> search(Map<String, String> searchParams) {
+		return new ArrayList<M>();
+	}
+	
 	public M get(ID id) {
-		E entity = repository.findById(id).get();
-		return mapper.mapEntityToModel(entity);
+		// TODO: Reflection helper
+		Optional<E> response = repository.findById(id);
+		if (!response.isPresent()) throw new GymServiceException("Entity not found (id: " + id + ").");
+		
+		return mapper.mapEntityToModel(response.get());
 	}
 	
 	public long count() {
 		return repository.count();
 	}
 	
-	public M save(M model) {
+	public M create(M model) {
+		if (model.getId() != null) throw new GymServiceException("Cannot create entity with id.");
+		
+		E entity = mapper.mapModelToEntity(model);
+		repository.save(entity);
+		return mapper.mapEntityToModel(entity);
+	}
+	
+	public M update(M model) {
+		if (model.getId() == null) throw new GymServiceException("Cannot update entity without id.");
+		
 		E entity = mapper.mapModelToEntity(model);
 		repository.save(entity);
 		return mapper.mapEntityToModel(entity);
 	}
 	
 	public boolean delete(M model) {
+		if (model.getId() == null) throw new GymServiceException("Cannot delete entity without id.");
+		
 		E entity = mapper.mapModelToEntity(model);
 		repository.delete(entity);
 		return true;
